VMO-6623 investigation 2
